feat(website): [T002150] website-db.ts split stage 2
