+++ docs/code-quality/repo-index.json
@@ -696,7 +696,7 @@
       "id": "website",
       "name": "Astro + Svelte website",
       "owner_agent": "bachelorprojekt-website",
-      "file_count": 1763,
+      "file_count": 1767,
       "files": [
         "website/.build-trigger",
         "website/.design-sync/NOTES.md",
@@ -1483,10 +1483,12 @@
         "website/src/lib/content-publish.ts",
         "website/src/lib/content-registry.test.ts",
         "website/src/lib/content-registry.ts",
+        "website/src/lib/content-store-db.ts",
         "website/src/lib/content.test.ts",
         "website/src/lib/content.ts",
         "website/src/lib/cors.test.ts",
         "website/src/lib/cors.ts",
+        "website/src/lib/custom-sections-db.ts",
         "website/src/lib/customer-crm-db.ensure.test.ts",
         "website/src/lib/customer-crm-db.test.ts",
         "website/src/lib/customer-crm-db.ts",
@@ -1680,6 +1682,7 @@
         "website/src/lib/poll-db.test.ts",
         "website/src/lib/poll-db.ts",
         "website/src/lib/popout.ts",
+        "website/src/lib/portal-tools-db.ts",
         "website/src/lib/project-attachments-db.ts",
         "website/src/lib/project-export-db.ts",
         "website/src/lib/project-portal-db.ts",
@@ -1846,6 +1849,7 @@
         "website/src/lib/tickets/transition.ts",
         "website/src/lib/tickets/triage-display.test.ts",
         "website/src/lib/tickets/triage-display.ts",
+        "website/src/lib/time-entries-db.ts",
         "website/src/lib/vat-id-validation.test.ts",
         "website/src/lib/vat-id-validation.ts",
         "website/src/lib/website-core-db.ts",
