+++ website/src/data/openspec-status.json
@@ -1413,6 +1413,12 @@
       "status": "plan_staged"
     }
   ],
+  "T002150": [
+    {
+      "slug": "t002150-website-db-split-stage-2",
+      "status": "plan_staged"
+    }
+  ],
   "TBD": [
     {
       "slug": "2026-06-21-secrets-deploy-automation",
