+++ website/src/lib/time-entries-db.ts
@@ -1,0 +1,514 @@
+import { pool } from './db-pool';
+import { initTicketsSchema } from './tickets-schema';
+
+let timeEntriesReady = false;
+
+export interface TimeEntry {
+  id: string;
+  projectId: string;
+  projectName: string;
+  taskId: string | null;
+  taskName: string | null;
+  description: string | null;
+  minutes: number;
+  billable: boolean;
+  rateCents: number;
+  leistungKey: string | null;
+  stripeInvoiceId: string | null;
+  entryDate: Date;
+  createdAt: Date;
+}
+
+async function initTimeEntriesTable(): Promise<void> {
+  if (timeEntriesReady) return;
+  await initTicketsSchema();
+  await pool.query(`
+    CREATE TABLE IF NOT EXISTS time_entries (
+      id                UUID        PRIMARY KEY DEFAULT gen_random_uuid(),
+      project_id        UUID        NOT NULL REFERENCES tickets.tickets(id) ON DELETE CASCADE,
+      task_id           UUID        REFERENCES tickets.tickets(id) ON DELETE SET NULL,
+      description       TEXT,
+      minutes           INTEGER     NOT NULL CHECK (minutes > 0),
+      billable          BOOLEAN     NOT NULL DEFAULT true,
+      rate_cents        INTEGER     NOT NULL DEFAULT 0,
+      stripe_invoice_id TEXT,
+      entry_date        DATE        NOT NULL DEFAULT CURRENT_DATE,
+      created_at        TIMESTAMPTZ NOT NULL DEFAULT now()
+    )
+  `);
+  await pool.query(`
+    CREATE INDEX IF NOT EXISTS time_entries_project_id_idx ON time_entries(project_id)
+  `);
+  await pool.query(`
+    ALTER TABLE time_entries ADD COLUMN IF NOT EXISTS rate_cents        INTEGER DEFAULT 0
+  `);
+  await pool.query(`
+    ALTER TABLE time_entries ADD COLUMN IF NOT EXISTS stripe_invoice_id TEXT
+  `);
+  await pool.query(`
+    ALTER TABLE time_entries ADD COLUMN IF NOT EXISTS leistung_key TEXT
+  `);
+  timeEntriesReady = true;
+}
+
+export async function getLastTimeEntryRate(): Promise<number> {
+  await initTimeEntriesTable();
+  const result = await pool.query(
+    `SELECT rate_cents FROM time_entries ORDER BY created_at DESC LIMIT 1`
+  );
+  return result.rows[0]?.rate_cents ?? 0;
+}
+
+export async function createTimeEntry(params: {
+  projectId: string;
+  taskId?: string;
+  description?: string;
+  minutes: number;
+  billable?: boolean;
+  rateCents?: number;
+  leistungKey?: string;
+  entryDate?: string;
+}): Promise<TimeEntry> {
+  await initTimeEntriesTable();
+  const result = await pool.query(
+    `INSERT INTO time_entries (project_id, task_id, description, minutes, billable, rate_cents, leistung_key, entry_date)
+     VALUES ($1, $2, $3, $4, $5, $6, $7, COALESCE($8::date, CURRENT_DATE))
+     RETURNING
+       id,
+       project_id        AS "projectId",
+       NULL::text        AS "projectName",
+       task_id           AS "taskId",
+       NULL::text        AS "taskName",
+       description,
+       minutes,
+       billable,
+       rate_cents        AS "rateCents",
+       leistung_key      AS "leistungKey",
+       stripe_invoice_id AS "stripeInvoiceId",
+       entry_date        AS "entryDate",
+       created_at        AS "createdAt"`,
+    [
+      params.projectId,
+      params.taskId ?? null,
+      params.description ?? null,
+      params.minutes,
+      params.billable ?? true,
+      params.rateCents ?? 0,
+      params.leistungKey ?? null,
+      params.entryDate ?? null,
+    ]
+  );
+  return result.rows[0] as TimeEntry;
+}
+
+export async function listTimeEntries(projectId: string): Promise<TimeEntry[]> {
+  await initTimeEntriesTable();
+  const result = await pool.query(
+    `SELECT te.id,
+            te.project_id        AS "projectId",
+            p.title              AS "projectName",
+            te.task_id           AS "taskId",
+            task.title           AS "taskName",
+            te.description,
+            te.minutes,
+            te.billable,
+            te.rate_cents        AS "rateCents",
+            te.stripe_invoice_id AS "stripeInvoiceId",
+            te.leistung_key      AS "leistungKey",
+            te.entry_date        AS "entryDate",
+            te.created_at        AS "createdAt"
+     FROM time_entries te
+     JOIN tickets.tickets p    ON p.id  = te.project_id
+     LEFT JOIN tickets.tickets task ON task.id = te.task_id
+     WHERE te.project_id = $1
+     ORDER BY te.entry_date DESC`,
+    [projectId]
+  );
+  return result.rows;
+}
+
+export async function listAllTimeEntries(params?: {
+  billable?: boolean;
+  since?: string;
+}): Promise<TimeEntry[]> {
+  await initTimeEntriesTable();
+  const result = await pool.query(
+    `SELECT te.id,
+            te.project_id        AS "projectId",
+            p.title              AS "projectName",
+            te.task_id           AS "taskId",
+            task.title           AS "taskName",
+            te.description,
+            te.minutes,
+            te.billable,
+            te.rate_cents        AS "rateCents",
+            te.stripe_invoice_id AS "stripeInvoiceId",
+            te.leistung_key      AS "leistungKey",
+            te.entry_date        AS "entryDate",
+            te.created_at        AS "createdAt"
+     FROM time_entries te
+     JOIN tickets.tickets p    ON p.id  = te.project_id
+     LEFT JOIN tickets.tickets task ON task.id = te.task_id
+     WHERE ($1::boolean IS NULL OR te.billable = $1)
+       AND ($2::date    IS NULL OR te.entry_date >= $2::date)
+     ORDER BY te.entry_date DESC`,
+    [params?.billable ?? null, params?.since ?? null]
+  );
+  return result.rows;
+}
+
+export async function setTimeEntryStripeInvoice(
+  ids: string[],
+  stripeInvoiceId: string | null
+): Promise<void> {
+  if (ids.length === 0) return;
+  await initTimeEntriesTable();
+  await pool.query(
+    `UPDATE time_entries SET stripe_invoice_id = $1 WHERE id = ANY($2::uuid[])`,
+    [stripeInvoiceId, ids]
+  );
+}
+
+export async function getTimeEntryIdsByInvoice(stripeInvoiceId: string): Promise<string[]> {
+  await initTimeEntriesTable();
+  const result = await pool.query<{ id: string }>(
+    `SELECT id FROM time_entries WHERE stripe_invoice_id = $1`,
+    [stripeInvoiceId]
+  );
+  return result.rows.map(r => r.id);
+}
+
+export interface UnbilledCustomerGroup {
+  customerId: string;
+  customerName: string;
+  customerEmail: string;
+  entries: Array<{
+    id: string;
+    projectId: string;
+    projectName: string;
+    description: string | null;
+    minutes: number;
+    rateCents: number;
+    entryDate: Date;
+  }>;
+}
+
+export async function getUnbilledBillableEntriesByCustomer(
+  year: number,
+  month: number  // 1-12
+): Promise<UnbilledCustomerGroup[]> {
+  await initTimeEntriesTable();
+  const startDate = `${year}-${String(month).padStart(2, '0')}-01`;
+  const endDate   = new Date(year, month, 0).toISOString().slice(0, 10);
+  const result = await pool.query(
+    `SELECT te.id,
+            te.project_id        AS "projectId",
+            p.title              AS "projectName",
+            te.description,
+            te.minutes,
+            te.rate_cents        AS "rateCents",
+            te.entry_date        AS "entryDate",
+            c.id                 AS "customerId",
+            c.name               AS "customerName",
+            c.email              AS "customerEmail"
+     FROM time_entries te
+     JOIN tickets.tickets p ON p.id = te.project_id
+     JOIN customers c ON c.id = p.customer_id
+     WHERE te.billable = true
+       AND te.stripe_invoice_id IS NULL
+       AND te.entry_date BETWEEN $1 AND $2
+       AND p.customer_id IS NOT NULL`,
+    [startDate, endDate]
+  );
+
+  const byCustomer = new Map<string, UnbilledCustomerGroup>();
+  for (const row of result.rows) {
+    if (!byCustomer.has(row.customerId)) {
+      byCustomer.set(row.customerId, {
+        customerId: row.customerId,
+        customerName: row.customerName,
+        customerEmail: row.customerEmail,
+        entries: [],
+      });
+    }
+    byCustomer.get(row.customerId)!.entries.push({
+      id: row.id,
+      projectId: row.projectId,
+      projectName: row.projectName,
+      description: row.description,
+      minutes: row.minutes,
+      rateCents: row.rateCents,
+      entryDate: row.entryDate,
+    });
+  }
+  return [...byCustomer.values()];
+}
+
+export async function deleteTimeEntry(id: string): Promise<void> {
+  await initTimeEntriesTable();
+  await pool.query('DELETE FROM time_entries WHERE id = $1', [id]);
+}
+
+export async function getProjectTotalMinutes(
+  projectId: string
+): Promise<{ total: number; billable: number }> {
+  await initTimeEntriesTable();
+  const result = await pool.query(
+    `SELECT
+       COALESCE(SUM(minutes), 0)::int                                          AS total,
+       COALESCE(SUM(CASE WHEN billable THEN minutes ELSE 0 END), 0)::int       AS billable
+     FROM time_entries
+     WHERE project_id = $1`,
+    [projectId]
+  );
+  return result.rows[0];
+}
+
+// ── Client Notes ──────────────────────────────────────────────────────────────
+
+export interface ClientNote {
+  id: string;
+  keycloakUserId: string;
+  content: string;
+  createdAt: Date;
+}
+
+async function initClientNotesTable(): Promise<void> {
+  await pool.query(`
+    CREATE TABLE IF NOT EXISTS client_notes (
+      id                UUID        PRIMARY KEY DEFAULT gen_random_uuid(),
+      keycloak_user_id  TEXT        NOT NULL,
+      content           TEXT        NOT NULL,
+      created_at        TIMESTAMPTZ NOT NULL DEFAULT now()
+    )
+  `);
+  await pool.query(`
+    CREATE INDEX IF NOT EXISTS client_notes_keycloak_user_id_idx ON client_notes(keycloak_user_id)
+  `);
+}
+
+export async function listClientNotes(keycloakUserId: string): Promise<ClientNote[]> {
+  await initClientNotesTable();
+  const result = await pool.query(
+    `SELECT id,
+            keycloak_user_id AS "keycloakUserId",
+            content,
+            created_at       AS "createdAt"
+     FROM client_notes
+     WHERE keycloak_user_id = $1
+     ORDER BY created_at DESC`,
+    [keycloakUserId]
+  );
+  return result.rows;
+}
+
+export async function createClientNote(keycloakUserId: string, content: string): Promise<ClientNote> {
+  await initClientNotesTable();
+  const result = await pool.query(
+    `INSERT INTO client_notes (keycloak_user_id, content)
+     VALUES ($1, $2)
+     RETURNING id,
+               keycloak_user_id AS "keycloakUserId",
+               content,
+               created_at       AS "createdAt"`,
+    [keycloakUserId, content]
+  );
+  return result.rows[0];
+}
+
+export async function deleteClientNote(id: string): Promise<void> {
+  await pool.query('DELETE FROM client_notes WHERE id = $1', [id]);
+}
+
+// ── Onboarding-Checkliste ─────────────────────────────────────────────────────
+
+export interface OnboardingItem {
+  id: string;
+  keycloakUserId: string;
+  label: string;
+  done: boolean;
+  sortOrder: number;
+}
+
+const DEFAULT_ONBOARDING_ITEMS = [
+  'Erstgespräch gebucht',
+  'Vertrag unterzeichnet',
+  'Nextcloud-Ordner erstellt',
+  'Mattermost-Kanal eingerichtet',
+  'Rechnungsadresse erfasst',
+  'Zugangsdaten versendet',
+];
+
+async function initOnboardingTable(): Promise<void> {
+  await pool.query(`
+    CREATE TABLE IF NOT EXISTS onboarding_items (
+      id                UUID    PRIMARY KEY DEFAULT gen_random_uuid(),
+      keycloak_user_id  TEXT    NOT NULL,
+      label             TEXT    NOT NULL,
+      done              BOOLEAN NOT NULL DEFAULT false,
+      sort_order        INTEGER NOT NULL DEFAULT 0
+    )
+  `);
+  await pool.query(`
+    CREATE INDEX IF NOT EXISTS onboarding_items_keycloak_user_id_idx ON onboarding_items(keycloak_user_id)
+  `);
+}
+
+export async function getOrCreateOnboardingChecklist(keycloakUserId: string): Promise<OnboardingItem[]> {
+  await initOnboardingTable();
+  const existing = await pool.query(
+    `SELECT id, keycloak_user_id AS "keycloakUserId", label, done, sort_order AS "sortOrder"
+     FROM onboarding_items
+     WHERE keycloak_user_id = $1
+     ORDER BY sort_order ASC`,
+    [keycloakUserId]
+  );
+  if (existing.rows.length > 0) return existing.rows;
+
+  // Seed defaults
+  for (let i = 0; i < DEFAULT_ONBOARDING_ITEMS.length; i++) {
+    await pool.query(
+      `INSERT INTO onboarding_items (keycloak_user_id, label, sort_order)
+       VALUES ($1, $2, $3)`,
+      [keycloakUserId, DEFAULT_ONBOARDING_ITEMS[i], i]
+    );
+  }
+
+  const seeded = await pool.query(
+    `SELECT id, keycloak_user_id AS "keycloakUserId", label, done, sort_order AS "sortOrder"
+     FROM onboarding_items
+     WHERE keycloak_user_id = $1
+     ORDER BY sort_order ASC`,
+    [keycloakUserId]
+  );
+  return seeded.rows;
+}
+
+export async function toggleOnboardingItem(id: string, done: boolean): Promise<void> {
+  await initOnboardingTable();
+  await pool.query('UPDATE onboarding_items SET done = $2 WHERE id = $1', [id, done]);
+}
+
+export async function resetOnboardingChecklist(keycloakUserId: string): Promise<void> {
+  await initOnboardingTable();
+  await pool.query(
+    'UPDATE onboarding_items SET done = false WHERE keycloak_user_id = $1',
+    [keycloakUserId]
+  );
+}
+
+// ── Follow-ups ────────────────────────────────────────────────────────────────
+
+export interface FollowUp {
+  id: string;
+  keycloakUserId: string | null;
+  clientName: string | null;
+  clientEmail: string | null;
+  reason: string;
+  dueDate: Date;
+  done: boolean;
+  createdAt: Date;
+}
+
+async function initFollowUpsTable(): Promise<void> {
+  await pool.query(`
+    CREATE TABLE IF NOT EXISTS follow_ups (
+      id                UUID        PRIMARY KEY DEFAULT gen_random_uuid(),
+      keycloak_user_id  TEXT,
+      client_name       TEXT,
+      client_email      TEXT,
+      reason            TEXT        NOT NULL,
+      due_date          DATE        NOT NULL,
+      done              BOOLEAN     NOT NULL DEFAULT false,
+      created_at        TIMESTAMPTZ NOT NULL DEFAULT now()
+    )
+  `);
+}
+
+export async function createFollowUp(params: {
+  keycloakUserId?: string;
+  clientName?: string;
+  clientEmail?: string;
+  reason: string;
+  dueDate: string;
+}): Promise<FollowUp> {
+  await initFollowUpsTable();
+  const result = await pool.query(
+    `INSERT INTO follow_ups (keycloak_user_id, client_name, client_email, reason, due_date)
+     VALUES ($1, $2, $3, $4, $5)
+     RETURNING id,
+               keycloak_user_id AS "keycloakUserId",
+               client_name      AS "clientName",
+               client_email     AS "clientEmail",
+               reason,
+               due_date         AS "dueDate",
+               done,
+               created_at       AS "createdAt"`,
+    [
+      params.keycloakUserId ?? null,
+      params.clientName ?? null,
+      params.clientEmail ?? null,
+      params.reason,
+      params.dueDate,
+    ]
+  );
+  return result.rows[0];
+}
+
+export async function listFollowUps(showDone = false): Promise<FollowUp[]> {
+  await initFollowUpsTable();
+  const result = await pool.query(
+    `SELECT id,
+            keycloak_user_id AS "keycloakUserId",
+            client_name      AS "clientName",
+            client_email     AS "clientEmail",
+            reason,
+            due_date         AS "dueDate",
+            done,
+            created_at       AS "createdAt"
+     FROM follow_ups
+     WHERE ($1 OR done = false)
+     ORDER BY due_date ASC`,
+    [showDone]
+  );
+  return result.rows;
+}
+
+export async function getDueFollowUps(): Promise<FollowUp[]> {
+  await initFollowUpsTable();
+  const result = await pool.query(
+    `SELECT id,
+            keycloak_user_id AS "keycloakUserId",
+            client_name      AS "clientName",
+            client_email     AS "clientEmail",
+            reason,
+            due_date         AS "dueDate",
+            done,
+            created_at       AS "createdAt"
+     FROM follow_ups
+     WHERE done = false AND due_date <= CURRENT_DATE
+     ORDER BY due_date ASC`
+  );
+  return result.rows;
+}
+
+export async function updateFollowUp(id: string, params: {
+  done?: boolean;
+  dueDate?: string;
+  reason?: string;
+}): Promise<void> {
+  const sets: string[] = [];
+  const values: unknown[] = [id];
+  let idx = 2;
+
+  if (params.done !== undefined) { sets.push(`done = $${idx}`); values.push(params.done); idx++; }
+  if (params.dueDate !== undefined) { sets.push(`due_date = $${idx}`); values.push(params.dueDate); idx++; }
+  if (params.reason !== undefined) { sets.push(`reason = $${idx}`); values.push(params.reason); idx++; }
+
+  if (sets.length === 0) return;
+  await pool.query(`UPDATE follow_ups SET ${sets.join(', ')} WHERE id = $1`, values);
+}
+
+export async function deleteFollowUp(id: string): Promise<void> {
+  await pool.query('DELETE FROM follow_ups WHERE id = $1', [id]);
+}
